+++ src/DataFilters/Grammar/Syntax/NotExpression.cs
@@ -5,7 +5,7 @@ namespace DataFilters.Grammar.Syntax
     /// <summary>
     /// An expression that negate wrapped inside
     /// </summary>
-    public class NotExpression : FilterExpression, IEquatable<NotExpression>
+    public sealed class NotExpression : FilterExpression, IEquatable<NotExpression>
     {
         /// <summary>
         /// Expression that the NOT logical is applied to
@@ -19,12 +19,16 @@ namespace DataFilters.Grammar.Syntax
         /// <exception cref="ArgumentNullException"><paramref name="innerExpression"/> is <c>null</c>.</exception>
         public NotExpression(FilterExpression innerExpression) => Expression = innerExpression ?? throw new ArgumentNullException(nameof(innerExpression));
 
+        ///<inheritdoc/>
         public bool Equals(NotExpression other) => Expression.Equals(other?.Expression);
 
+        ///<inheritdoc/>
         public override bool Equals(object obj) => Equals(obj as NotExpression);
 
+        ///<inheritdoc/>
         public override int GetHashCode() => Expression.GetHashCode();
 
+        ///<inheritdoc/>
         public override string ToString() => $"{GetType().Name} : Expression ({Expression.GetType().Name}) -> {Expression}";
     }
 }