+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0] / 2021-05-02
+- Introduces [`PropertyNameResolutionStrategy`](/src/DataFilters/Casing/PropertyNameResolutionStrategy.cs) as extension point to configure the way to lookup 
+for a corresponding property ([#8](https://github.com/candoumbe/datafilters/issues/8))
+- Added `ToFilter<T>(string, PropertyNameResolutionStrategy)` overload
+- Added `ToSort<T>(string, PropertyNameResolutionStrategy)` overload
+
 ## [0.4.1] / 2021-04-28
 - Fixes `FileNotFoundException` when calling `StringExtensions.ToFilter<T>` method after a fresh install.
 
@@ -27,7 +33,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.3.0] / 2020-12-19
 - Added support for filtering over complex type collections by introducing sub property syntax
 - Moved `FilterExtensions.ToFilter<T>(this StringSegment)` method to `StringExtensions.ToFilter<T>(this StringSegment)` [BREAKING]
-- Moved `FilterExtensions.ToFilter<T>(this string)` method to `StringExtensions.ToFilter<T>(this string)` [BREAKING]
+- Moved `FilterExtensions.ToFilter<T>(this string)
+- 
+` method to `StringExtensions.ToFilter<T>(this string)` [BREAKING]
 - Changed `ConstantExpression` to `ConstantValueExpression` [BREAKING]
 - Changed subproperty syntax from `property.subproperty` method to `property["subproperty"]` [BREAKING]
 - Fixed parsing [`RangeExpression`](src/DataFilters/Grammar/Syntax/RangeExpression.cs) with datetime values
@@ -41,7 +49,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for "equals" operator on collections
 - Added support for "contains" operator on collections.
 
-[Unreleased]: https://github.com/candoumbe/DataFilters.git/compare/0.4.1...HEAD
+[Unreleased]: https://github.com/candoumbe/DataFilters.git/compare/0.5.0...HEAD
+[0.5.0]: https://github.com/candoumbe/DataFilters.git/compare/0.4.1...0.5.0
 [0.4.1]: https://github.com/candoumbe/DataFilters.git/compare/0.4.0...0.4.1
 [0.4.0]: https://github.com/candoumbe/DataFilters.git/compare/0.3.2...0.4.0
 [0.3.2]: https://github.com/candoumbe/DataFilters.git/compare/0.3.1...0.3.2